feat: bake transactions into Docker images
There's been some confusion around the state of the translations in the repo as well as within different builds (docker, pypi, official releases,...).

In this PR, we assume that "compiled" or redundant files that can be generated don't belong in the repository, and should be baked as part of the build process.

Yet to TO DO to get this out of DRAFT:
- make sure .json files are generated as part of the pypi/ASF build process
- read through the docs and make sure it's clear how to build things if you're not using docker or pypi